Test Cases With Emails
* Debug messages and some IFs.

+++ tests/assets/cases/scripts/clearemails.sh
@@ -1,3 +1,5 @@
 #!/bin/bash
 #
-find "${TRAVISCI_EMAIL_TMP}/fake-mailbox" -type f | xargs sudo rm -fv;
+if [ -d "${TRAVISCI_EMAIL_TMP}/fake-mailbox" ]; then
+	find "${TRAVISCI_EMAIL_TMP}/fake-mailbox" -type f | xargs sudo rm -fv;
+fi;
